saori-revive/requirements.txt
discord.py
lavalink
PyNaCl
aiohttp
asyncio
